Refactor test helpers into helper files
Following 5caa083 this commit extracts more helper functionality from test suites into `helpers_test.go` files. Specifically, it does this for the test suites of:

- internal/cli/parse.go
- internal/replacer/replacer.go
