+++ internal/replacer/replacer_test.go
@@ -7,14 +7,6 @@ import (
 	"github.com/ericcornelissen/wordrow/internal/wordmaps"
 )
 
-func reportIncorrectReplacement(t *testing.T, expected, actual string) {
-	t.Helper()
-	t.Errorf(`Replacement did not work as intended
-		expected : '%s'
-		got      : '%s'
-	`, expected, actual)
-}
-
 func TestReplaceEmptyString(t *testing.T) {
 	var wm wordmaps.WordMap
